sgcmapp/src/app/component/convenio/convenio-list/convenio-list.component.ts
import { CommonModule } from '@angular/common';
import { Component } from '@angular/core';
import { RouterLink } from '@angular/router';
import { Convenio } from '../../../model/convenio';
import { BarraComandosComponent } from '../../barra-comandos/barra-comandos.component';
import { IList } from '../../i-list';
import { ConvenioService } from '../../../service/convenio.service';
import { AlertaService } from '../../../service/alerta.service';
import { ETipoAlerta } from '../../../model/e-tipo-alerta';

@Component({
  selector: 'app-convenio-list',
  standalone: true,
  imports: [CommonModule, BarraComandosComponent, RouterLink],
  templateUrl: './convenio-list.component.html',
  styles: ``
})
export class ConvenioListComponent implements IList<Convenio> {

  constructor(
    private servico: ConvenioService,
    private servicoAlerta: AlertaService
  ) { }

  ngOnInit(): void {
    this.get();
  }

  registros: Convenio[] = Array<Convenio>();

  get(termoBusca?: string): void {
    this.servico.get(termoBusca).subscribe({
      next: (resposta: Convenio[]) => {
        this.registros = resposta;
      }
    });
  }

  delete(id: number): void {
    if (confirm('Confirma a exclusão do convênio?')) {
      this.servico.delete(id).subscribe({
        complete: () => {
          this.get();
          this.servicoAlerta.enviarAlerta({
            tipo: ETipoAlerta.SUCESSO,
            mensagem: "Convênio excluído com sucesso!"
          });
        }
      });
    }
  }

}
